fix: Correctly delete categories and related offers
Fixes an issue where deleting a category didn't properly
cascade the deletion to related offers.  Also, ensures that
selected attributes are sent to the API when creating
categories.
The GameController's destroy method now iterates through
the categories associated with a game to delete them
individually. This is necessary to trigger the deleting event
on the Category model, which handles the deletion of related
offers.

The CategoryCreate.vue component now uses `selectedAttributes`
instead of `attributes` when creating a new category of type 1.